Create getrevueexporter.php
as generated by ChatGPT-3

Create an internationalized wordpress plugin named "getrevueexporter" authored by "politbuero-kampagnen.ch". The plugin could be configured to access a specific getrevue account, then show all the available newsletters for export, allow the user to choose which newsletters to export and whether these should be saved as draft posts or published posts. All instructions should be in the code as comment, i.e. how to create an api-key for getrevue.co